bugfixes und scripte

+++ src/domain/AbstractLieferant.java
@@ -27,7 +27,7 @@ public abstract class AbstractLieferant extends Thread implements Serializable{
 	}
 	
 	public void startLieferant() throws InterruptedException{
-		long abladeZeit = ((new Random().nextLong())%1000)+1000;
+		long abladeZeit = (new Random().nextInt()%1000)+1000L;
 		for(int i = 0; i< anzahl ; i++){
 			Thread.sleep(abladeZeit);
 			zutatAbladen(zutatTyp);
